Avoid placing Screen as elements' parent.
Since screen is already accessible in every Element via `#screen`,
and since Screen does not implement the interface of an `Element`
yet many places where #parent is called expect an Element, Screen
should not be listed as any one Element's parent.

However, for now, attaching an Element to a Screen is still done
via `#append` (which is used for adding an element to an element
as well), and Screen keeps a list of its @children so that they
can be rendered.

So adding/removing children elements works as usual; the only step
skipped at this point is the literal setting of:

element.parent = parent_el

in cases where parent_el is indeed Screen.

This approach seems to work and should ideally allow us to define
that the type of `parent` is `Element` rather than `Node`.