Second step of removing Screen as parent
The changes from this commit should be improved altogether actually,
so that checks in style of (parent or screen) are never even made,
but simply default to 0 or so when parent is nil.

Also, some parts have remained using @parent instead of (parent or
screen). I believe this is correct since a Screen in those positions
is of no use/makes no difference. But it is possible that this will
be proven false over time and will need changing.