Improvement for alignment behavior
When {center} or {right} is found in widget content, and parse_tags==true, this
will align the text to center or right. As a side effect, 100% of the width will
be padded with empty chars where there is no content.

Previously, this was different than {left} align, which wasn't aligned in any
particular way, and thus also not padded.

With this change, {left} keeps the text on the left side as it was before, but
as a side effect the whole row is padded (analogous and opposite of {right}).