Streamline meaning of attach/detach
In Blessed, Screen is a top-level parent of every Widget. This forced Screens and
Widgets to share some part of API which is not otherwise applicable.

In Crysterm, Screen is not a top-level parent (only Widgets are), and screen is
accessible by calling `#screen` on a widget. So:

Attach and detach operations now uniformly refer to attaching/detaching
widgets from a screen.

Usually this is not called manually but happens as a side-effect of
assigning widgets' parent elements.

Note: attach/detach do not add a widget to list of particular screen's children
elements -- they only set its property #screen.