Focus improvement
Previously, if no Widget was in focus yet, calling `screen.focus_next` was not
working because it required widget to be focused beforehand.

Now, it simply focuses on the first focusable widget, avoiding the need to
differentiate between cases when at least one widget was focused previously
and when it wasn't.