Remove left/top/right/bottom aliases
These were aliases for the relative distances (rleft, rtop, rright, bottom).

But we already have the 'r' ones, plus 'a' ones for absolute distances
from Point 0,0 of Screen, and also 'i' ones for the distances of the
"inner" content (i.e. after border and padding are added).

So having left/top/right/bottom is too confusing since one always has to
remember which of the three it is an alias for.

So, remove those aliases and expect users to use r/a/i explicitly.