Improve focus positioning in checkboxes
In checkboxes/radio buttons, when widget is focused, dwthe hardware cursor
shows up in place of the checkmark.

A deficiency inherited from blessed was that this did not take into account
inner dimensions (border and/or padding currently).

This has now been fixed. Still, the work remains to support or verify
artificial cursor support as well.