Improve initial widget focus
When a screen renders for the first time and a particular widget is
selected to have focus, emit Event::Focus on it.

This makes widget's behavior/visuals behave identically as when
(re)focused later via focus switching.

Most notable location was in e.g. checkboxes/radiobuttons which
would not be marked as having focus on screen's first render,
until focus was cycled with [Shift+]Tabs.